modified:   .github/workflows/ci.yml 	modified:   Makefile 	deleted:    assets/images/swagger.png 	deleted:    docs/src/mer.wsd 	modified:   pom.xml 	new file:   src/main/java/com/alfa/devops/DemoApplication.java 	new file:   src/main/java/com/alfa/devops/web/HelloControler.java 	deleted:    src/main/java_ping/Application.java 	deleted:    src/main/java_ping/controller/PingController.Java 	modified:   src/main/resources/application.properties 	deleted:    src/test/PingControllerTest.java 	new file:   src/test/java/com/alfa/devops/web/HelloControllerTest.java
