+++ .github/workflows/ci.yml
@@ -4,19 +4,25 @@ on:
     branches: ["feature/*"]
   pull_request:
     branches: ["main"]
+
 jobs:
   build:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+
       - name: Set up JDK 21
         uses: actions/setup-java@v4
         with:
           java-version: '21'
           distribution: 'temurin'
           cache: maven
-      - run: chmod +x mvnw
-      - name: Test
-        run: make test
+
+      - name: Ensure mvnw is executable
+        run: if [ -f mvnw ]; then chmod +x mvnw; fi
+
+      - name: Testes
+        run: ./mvnw -B -ntp test || mvn -B -ntp test
+
       - name: Build
-        run: make build
+        run: ./mvnw -B -ntp clean package || mvn -B -ntp clean package
