+++ src/main/java/com/alfa/devops/DemoApplication.java
@@ -1,0 +1,11 @@
+package com.alfa.devops;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DemoApplication {
+  public static void main(String[] args) {
+    SpringApplication.run(DemoApplication.class, args);
+  }
+}
