+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
 server.port=8080
-spring.application.name=java-ping
-springdoc.swagger-ui.path=/
+# mensagem default; pode ser sobrescrita via env MENSAGEM
+app.message=Aplicação de demo para pipeline
