+++ src/test/java/com/alfa/devops/web/HelloControllerTest.java
@@ -1,0 +1,14 @@
+package com.alfa.devops.web;
+
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+
+class HelloControllerTest {
+
+  @Test
+  void pingDeveRetornarPong() {
+    HelloController c = new HelloController();
+    String resp = c.ping();
+    assertEquals("pong", resp);
+  }
+}
